Refactor structures to support tracking locations
This commit refactors several structures (Location, Function, Identifier, etc.) in the `src/location.rs`, `src/parser.rs` and `src/ast.rs` files to support tracking the beginning and end positions of parsed constructs. This is done by introducing the `start_offset` and `end_offset` fields in the `Location` struct and using the `LocatedSpan` object from the `nom_locate` dependency in the `parse_` functions. This change is necessary to provide precise error messages.