Improve AST and parser
The parser now returns a LilError which is more in line with the
rest of the application.

In general the error handling is much improved with code highlights
for single line and multiline errors.

The for and if statemetns can now have optional parts, like the
for not requireing a initialization or updation and ifs can have
an optional else branch.