Introduce register allocation, computation of live intervals, and logging
This commit introduces computation of live intervals in LIR functions, logging enabled in different parts of JIT compilation, and allocation of registers in LIR function compilation. It refactors the LIR compiler with a new structure for maintaining LIR functions and the JIT compiler by implementing the logger. Also, a new debug display trait was implemented for clearer output of debug prints.