Refactor register definition and compiler for better architecture flexibility
This commit restructures the register definitions module and expands the functionalities of the 'jit' compiler. Specifically, the 'reg_def' module was renamed to 'arch_def' and moved to the 'jit' folder. This enhances usability and allows for better flexibility in support different processor architectures. Compiler capabilities have been extended to make use of the new architecture -specific definitions, improving code efficiency and readability. This modification is critical for the compatibility and performance of the JIT compiler with multiple architectures. A new file 'arm64.rs' has been added for ARM64 specific register definitions. As extension of the compiler functionality, more detailed operations, including instructions for binary operations, are now included in the 'compiler.rs' file. The refactoring also resulted in minor modifications in 'reg_alloc.rs', 'jit.rs' and 'funcinfo.rs' to align them with the new structure and enhanced compiler functionalities.