Integrate support for JIT compile time function stubs
This commit imports the newly added `compile_stub` function to the compiler logic in `src/jit/compiler.rs`. When the compiler fails to find a compiled version of a function in the JIT data structs, it now calls `compile_stub` to dynamically generate a stub function. This function is implemented in assembly in the new file `src/jit/stub.rs`, using `global_asm!`. The stub saves the link register to the stack, calls the JIT compiler, then restores the link register and returns to the caller.

To allow the stub to call the JIT compiler, a function `sink` was added to `src/jit.rs`, along with a mutable static reference to the current JIT instance, `JIT_REF`. Because static mutable references are unsafe, usage of `JIT_REF` is acquired and checked within a newly added unsafe block within the `run_jit` function.

Finally, some minor refactoring in `src/lib.rs` and `src/main.rs` was needed to follow the new structure.