Implement input argument loading to JIT compiler
This commit implements the handling of input arguments in the JIT compiler. By introducing a new `InputArgLoad` LIR, it is now possible to reference input arguments and load them into registers when required.
Also, the argument load handling is added in the `live_interval.rs` to manage callee and caller saved registers properly.

Further, it fixes a small bug in example `problem_code.lil`. Now the function `test(2)` gets called before `return test(0)` statement.

Lastly, a necessary import statement is added to `compiler.rs` that upgrades the quality of our compiler to assert potential PANICs which in result makes it more robust.

In summary, these changes enhance the languages' feature set and robustness of the JIT compiler.