Adjusted function call implementation and refactor some codes
This commit refactored some parts of the code which involve function call implementations. A `mov_arbitrary_imm` method was added to efficiently load an arbitrary immediate value into a register. The function call method in `src/jit/compiler.rs` was updated to use this method, covering all possible function locations albeit being inefficient for short offset. A similar adjustment was also made for the call patch method in `src/jit.rs.`
Other changes include minor adjustment in the order of imports and the removal of unused pieces of codes.
The changes will result in more modern and efficient codes.