Add Dockerfile to setup development environment
A new Dockerfile has been added to the .devcontainer directory to streamline the development environment setup. This Dockerfile begins with an ARM64 Rust base image, installs libclang-dev as a system dependency, and sets /workspace as the working directory for convenience. Additionally, the application's existing Cargo manifests are copied over, with the expectation that `cargo fetch` will be run to populate the build cache. Lastly, the rest of the application is copied over.