Update logging and rectify recompilation warning in 'jit.rs'
This commit enhances the logging in 'jit.rs' by utilizing an 'info' log level instead of 'warn' whenever a function gets recompiled. It also corrects a minor issue in 'reg_alloc/reg_mapping.rs' by changing the unnecessary key match to a value match in the 'number_of_spills' method. It amends the unused constant and function in 'jit/codegendata.rs' by adding '#[allow(dead_code)]' and '#[allow(unused)]' attributes. The call to extend_memory was also commented out due to being temporarily disabled. The aim of these changes is to improve the verbosity of debug logs and to mitigate unnecessary recompilation of functions.