Refactor Docker setup for 'lil_jit' service
The Docker configuration for the 'lil_jit' service was adjusted to simplify the launch process, increase efficiency and fit a more idiomatic Docker usage. We removed the 'docker-compose.yaml', as the required environment configuration got too simple to necessitate it. This simplicity was achieved by moving build instructions and the running command to the 'Dockerfile'.
Additionally, a '.dockerignore' file was created to avoid including the 'target' directory in the Docker context, optimizing image build time. A 'lil_jit.sh' shell script was added to automate Docker image build and run steps with application arguments. These changes ease the debugging process and deployment.