Remove unused DOCKERFILE_PATH variable from lil_jit.sh
Removed the unnecessary variable DOCKERFILE_PATH from the shell script that runs the Docker image. Updated the script to directly use the Dockerfile path './Dockerfile' instead of referring it from a variable which was adding unnecessary complexity to the code. This change simplifies the script and makes it more robust by removing dependency on an extra variable.