Modify README.md for clarity and correctness
Improved the instructions regarding the use of Docker in Readme.md. Placed `lil_jit` in inline code blocks to signify it as a command or executable. Corrected a typographical error in providing information about Docker requires an emulator for non ARM64 architectures. The phrase originally read "build run" but it should be "build and run". For the running instruction, efficiency was improved by condensing the details into one complete sentence removing the unnecessary line break.