Extended JIT compiler's function argument handling to include constants
In this commit, the JIT compiler is now able to handle function arguments that are constants in addition to registers. This was implemented in the `Call` function by extending it to handle `LirOperand::Constant` cases. The stack space calculation for arguments was also adapted to accommodate constants, making use of the `mov_arbitrary_imm` function to move constant values into registers. This change provides better optimization opportunities when functions are called with constant arguments.