Add constant single assignment optimization to LIR
Introduced constant single assignment optimization in the Low-Level Intermediate Representation (LIR) stage of JIT compiler. New helper functions 'calc_constant' and 'new_operand_of' have been added in the optimizer for constant calculation and operand replacement. With these changes, if a variable is assigned only once, and the value is a constant, the variable assignment can be replaced with the constant directly, further enhancing the performance of the JIT compiler. Also, added related test cases.