Add constant optimization setting in JIT compiler
Introduced an atomic boolean for constant optimization in the Just-in-Time (JIT) compiler settings, encapsulated in a 'JitSettings' struct. Added also a command-line argument 'optimize_constants' for the users to enable or disable this feature as needed. If 'optimize_constants' is enabled, the JIT would optimize the constant computations in the Low-Level Intermediate Representation (LIR) stage, potentially enhancing the execution speed.